SNOW-181550 Don't use the RST note admonition
Description
Don't use the RST note admonition.
On GitHub, after the HTML is generated from an RST file, CSS classes are stripped out of the HTML. This means that in the
resulting HTML for a note, there isn't anything in the HTML that identifies the content as a note, and the browser just displays
the note as a plain block of text. For the reader, it's not clear where the note ends.